Rename ro scope_exit
