+++ test/unit_test/CMakeLists.txt
@@ -11,7 +11,7 @@ add_executable(vw-unit-test.out
   options_boost_po_test.cc
   options_test.cc
   prediction_test.cc
-  scope_guard_test.cc
+  scope_exit_test.cc
   slates_parser_test.cc
   slates_test.cc
   stable_unique_tests.cc
