+++ test/unit_test/unit_test.vcxproj
@@ -32,7 +32,7 @@
     <ClCompile Include="options_boost_po_test.cc" />
     <ClCompile Include="options_test.cc" />
     <ClCompile Include="prediction_test.cc" />
-    <ClCompile Include="scope_guard_test.cc" />
+    <ClCompile Include="scope_exit_test.cc" />
     <ClCompile Include="slates_parser_test.cc" />
     <ClCompile Include="slates_test.cc" />
     <ClCompile Include="stable_unique_tests.cc" />
